[ENH] History and reorganization
