+++ peakdet/editor.py
@@ -3,23 +3,10 @@
 import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib.widgets import SpanSelector
-from peakdet import physio
+from peakdet import physio, utils
 
 
-def edit_physio(data):
-    """
-    Returns interactive physio editor for `data`
-
-    Parameters
-    ----------
-    data : Physio_like
-        Physiological data to be edited
-    """
-
-    return PhysioEditor(data)
-
-
-class PhysioEditor():
+class _PhysioEditor():
     """ Class for editing physiological data
 
     Parameters
@@ -28,30 +15,34 @@ class PhysioEditor():
         Physiological data to be edited
     """
 
-    def __init__(self, data, _debug=False, _xlim=None):
-        if not isinstance(data, physio.PeakFinder):
-            raise TypeError('Input must be a sub-class of peakdet.PeakFinder')
+    def __init__(self, data):
+        if not isinstance(data, physio.Physio):
+            raise TypeError('Input must be a Physio instance')
         # save reference to data, time for plotting
         self.data = data
         self.time = np.arange(0, data.size / data.fs, 1. / data.fs)
-        self.defxlim = _xlim
         self.plot = False
+
         # make main plot objects
         self.fig, self.ax = plt.subplots(nrows=1, ncols=1, tight_layout=True)
         self.fig.canvas.mpl_connect('scroll_event', self.on_wheel)
         self.fig.canvas.mpl_connect('key_press_event', self.on_key)
-        self.span = SpanSelector(ax=self.ax, onselect=self.on_span, button=1,
-                                 direction='horizontal', useblit=True,
-                                 rectprops=dict(facecolor='red', alpha=0.3))
+        self.span1 = SpanSelector(self.ax, self.on_reject, 'horizontal',
+                                  button=1, useblit=True,
+                                  rectprops=dict(facecolor='red', alpha=0.3))
+        self.span2 = SpanSelector(self.ax, self.on_delete, 'horizontal',
+                                  button=3, useblit=True,
+                                  rectprops=dict(facecolor='blue', alpha=0.3))
         self.plot_signals()
 
     def plot_signals(self):
         """ Clears axes and plots data / peaks / troughs """
+        # don't reset x-/y-axis zooms on replot
         if self.plot:
             xlim, ylim = self.ax.get_xlim(), self.ax.get_ylim()
         else:
             self.plot = True
-            xlim, ylim = (-5, self.defxlim), (None, None)
+            xlim, ylim = (-5, None), (None, None)
 
         self.ax.clear()
         self.ax.plot(self.time, self.data, 'b',
@@ -68,15 +59,37 @@ class PhysioEditor():
         self.ax.set_xlim(lim[0] + move, lim[1] + move)
         self.fig.canvas.draw()
 
-    def on_span(self, xmin, xmax):
+    def on_reject(self, xmin, xmax):
         """ Manually removes physio peaks on span select """
-        imin, imax = np.searchsorted(self.data.peaks,
-                                     np.searchsorted(self.time, (xmin, xmax)))
-        remove = np.arange(imin, imax, dtype=int)
-        self.last_removed = self.data.peaks[remove]
-        self.data._reject = np.append(self.data._reject,
-                                      self.last_removed)
-        self.data._find_troughs()
+        rej = np.arange(*np.searchsorted(self.data.peaks,
+                                         np.searchsorted(self.time,
+                                                         (xmin, xmax))),
+                        dtype=int)
+        if len(self.data.peaks[rej]) == 0:
+            return
+        self.data._history.append(('manual_reject',
+                                   self.data.peaks[rej].tolist()))
+        self.data._metadata.reject = np.append(self.data._metadata.reject,
+                                               self.data.peaks[rej])
+        self.data._metadata.troughs = utils.check_troughs(self.data,
+                                                            self.data.peaks,
+                                                            self.data.troughs)
+        self.plot_signals()
+
+    def on_delete(self, xmin, xmax):
+        """ Manually delete physio peaks on span select """
+        rej = np.arange(*np.searchsorted(self.data.peaks,
+                                         np.searchsorted(self.time,
+                                                         (xmin, xmax))),
+                        dtype=int)
+        if len(self.data.peaks[rej]) == 0:
+            return
+        self.data._history.append(('manual_delete',
+                                   self.data.peaks[rej].tolist()))
+        self.data._metadata.peaks = np.delete(self.data._metadata.peaks, rej)
+        self.data._metadata.troughs = utils.check_troughs(self.data,
+                                                            self.data.peaks,
+                                                            self.data.troughs)
         self.plot_signals()
 
     def on_key(self, event):
@@ -88,12 +101,28 @@ class PhysioEditor():
 
     def undo(self):
         """ Resets last span select peak removal """
-        if self.last_removed is None:
+        # check if last history entry was a manual reject / delete
+        if len(self.data.history) == 0:
             return
-        self.data._reject = np.setdiff1d(self.data._reject,
-                                         self.last_removed)
-        self.data._find_troughs()
-        self.last_removed = None
+        elif self.data.history[-1][0] not in physio._ACCEPTED_HISTORY:
+            return
+        # pop off last edit
+        func, peaks = self.data.history.pop()
+        if func == 'manual_reject':
+            self.data._metadata.reject = np.setdiff1d(
+                self.data._metadata.reject, peaks
+            )
+        elif func == 'manual_delete':
+            self.data._metadata.peaks = np.insert(
+                self.data._metadata.peaks,
+                np.searchsorted(self.data._metadata.peaks, peaks),
+                peaks
+            )
+        else:
+            raise ValueError('This should never happen.')
+        self.data._metadata.troughs = utils.check_troughs(self.data,
+                                                            self.data.peaks,
+                                                            self.data.troughs)
         self.plot_signals()
 
     def quit(self):
